backend/requirements.txt
# Core Backend
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.20
pydantic==2.10.3
pydantic-settings==2.6.1

# Database
psycopg2-binary==2.9.10
sqlalchemy==2.0.36
alembic==1.14.0
pgvector==0.3.6

# Redis
redis==5.2.1
hiredis==3.0.0

# Google Cloud Storage
google-cloud-storage==2.19.0
google-auth==2.37.0

# Document Processing
PyMuPDF==1.25.1
Pillow==11.0.0
pytesseract==0.3.13
dl-translate==0.3.0
indic-nlp-library==0.92
python-docx==1.1.2

# LangChain & AI
langchain==0.3.13
langchain-community==0.3.13
langchain-core==0.3.28
langchain-experimental==0.3.3
langchain-text-splitters==0.3.4
langchain-ollama==0.2.2
langchain-google-genai==2.0.8
ollama==0.4.5

# Google Gemini (for local development only)
# Only installed when USE_GEMINI_FOR_DEV=true
# Safe to keep in requirements as it's optional
google-generativeai==0.8.3

# Neo4j (for knowledge graph storage)
neo4j==5.27.0
langchain-neo4j==0.2.1

# Auth & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dateutil==2.9.0.post0
bcrypt==4.2.1

# Utilities
python-dotenv==1.0.1
httpx==0.28.1
aiofiles==24.1.0
requests==2.32.3

# Additional dependencies for production
gunicorn==23.0.0  # Production WSGI server (alternative to uvicorn)
sentry-sdk==2.19.2  # Error tracking (optional)

